style: cve: add maxlength to description and notes
Maximum length validation for the description and notes field for CVE
and AVG is done on the backend. As of now, the application truncates any
description that exceeds the LENGTH definition. Although this is the
right way to ensure no excess data is given, descriptions and notes may
end truncated with prior notice to the user, which may be undesirable.

Add a maxlength attribute to the textarea fields, as well as a validator
for the cve and avg form and form template.