style: Add max-length constants to CVE/AVG models
Currently, CVE and AVG models contain description and notes fields with
a maximum length defined as a magical number. Given that these values
will be used throughout the codebase, moving them to a constant within
the same class will avoid repeating the magic number everywhere else.

Add a constant for the CVEGroup and CVE models that describe the maximum
length of the notes and description fields.