app:view:edit: fix wrong CVE list parsing
When editing AVG's, the view handler splits the data from the form using
'\r\n'. This allows for the creation of spureous CVE values if two cve's
are input using a space, instead of a new line and carriage return.

Split the CVE field using multiline_to_list. This is the method used for
the add views, and it handles this case properly.