app: view: show: truncate summary length
The form under bugs.archlinux.org has a maximum field length of 200. If
the length of the summary field computed is larger than this, the
propulated form will fail to submit, which requires the user (submitter)
to manually edit the summary field to have a shorter value.

Add a check to the get_bug_data function to use an alternative summary
if this length is exceeded.