advosry: fix change type accordingly if the group issues change
Change the advisory issue type accordingly when adding or removing
CVE entries from a group that has already scheduled multiple advisories
because its multi package.

The issue was a missing group-by on the pkgname which resulted in only
selecting one of the multiple advisories that got scheduled and the
others kept its old advisory issue types when it should have been
changed.

This fixes #100