OTP authentication draft
This introduces a WIP OTP authentication login mechanism for the
tracker. This introduces a hack, and proves a possible implementation
error.

The views themselves can not change the `is_authorized` attribute. Only
the `load_user` function in `user.py` does this as our only check is on
`current_user`. This function is only creating `current_user` after the
initial `login_user(user)` call done after successfully verification.

Any calls to `login_user(user)` will authenticate the session and no
partial session can be created. To get around this we check if OTP
authentication is enabled, then create a session variable `two_factor`
which is checked inside the view and form to prompt further
authentication.

Signed-off-by: Morten Linderud <morten@linderud.pw>