tracker/view/login: rewire redir index->index_vuln
The old behavior that followed a login attempt was to redirect to the
index, which used to be the index (equivalent to index_vulnerable). With
the new security splash page, this behavior breaks. While it is still up
to discussion whether the after-login behavior should be to redirect to
the security splash page instead of index_vulnerable, for now we still
need to keep the existing behavior. Rewire the redirects on from
tracker.index to tracker.index_vulnerable on the login views. Finally,
update test_login to reflect this behavior.

Signed-off-by: Santiago Torres <santiago@archlinux.org>