cve: fix is_modified check when severity or type gets changed
Changing the severity or the type will lead to some complex queries
involved to update the severity or type of groups referencing this
issue. In case this happens, some implicit commits may mark the model
as not modified, therefor explicitly check for changed severity or type
and force update the change state.