feature(sso): implement optional SSO (openid) authentication
This implements an optional new authentication method via openid connect
to be used with Keycloak on the live Arch Linux instance. SSO is
configured in the security trackers configuration file which requires
the metadata_url (openid-configuration), client_id, client_secret and
the group mappings. The groups should be part of the token. When SSO is
enabled existing accounts are mapped to locally existing accounts on
email address, once the user is created in the database the next
authentication uses the idp_id to match the user. Users who authenticate
and are not part of any security tracker group are not logged in but can
view the security tracker.

Co-Authored-By: Jelle van der Waa <jelle@archlinux.org>
Co-Authored-By: Levente Polyak <anthraxx@archlinux.org>