fix(db): automatically stamp the alembic migrate on db init
If a new database is initialized the alembic meta data is not
automatically stamped to the latest head. This leads to issues when
trying to create migration scripts. As a fix we automatically stamp the
alembic meta data to the current migration scripts head if the database
did not exist before.

This however does not automatically fix existing databases that lack the
alembic meta data stamp. Those should only affect some dev systems and
instead of carrying around some patch paths, we expect devs to simply
call ./trackerctl db stamp head

Fixes #182