Add group creation date to issues and group json
Through the json interface currently there is no way to find when a
group was created. If an app for example uses /issues/vulnerable.json
and needs this information it currently additionally has to download the
group html for each group and parse the "Created date" from the html.

This MR adds group.created as 'date' to the issues and group json so for
example downloading /issues/vulnerable.json would include this
information and no additional download for each group is needed.

It uses Y-M-d date format same as is used for advisories creation date
in package json.